Add files via upload
This is the application of the Siamese Neural Network (SNN for short) on the Mnist dataset. The goal of this project is instead of training the model to identify the number by what it is, numbers are identified by their similarity to another number. For example, if two 6's are sent into the Model, a score of Similar should be expected but if two different numbers are sent, a score of different should be expected. There was some variance in the train_loss and train_acc but the val_loss and val_acc had low variance and low bias. I highly doubt that there was a leak in the data as the model performed well on the test section.